chord-part-2/chord-part-2/requirements.txt
boto3
msgpack-rpc-python